+++ pytest_adaptavist/_atm_configuration.py
@@ -1,4 +1,5 @@
 """Read config parameters."""
+
 import json
 import os
 from typing import Any, Optional
